Made some fixes about names and methods
